Added password sharing capabilities

+++ src/lib/Controller/Api/ShareApiController.php
@@ -8,6 +8,7 @@
 
 namespace OCA\Passwords\Controller\Api;
 
+use OCA\Passwords\Db\Password;
 use OCA\Passwords\Db\PasswordRevision;
 use OCA\Passwords\Db\ShareRevision;
 use OCA\Passwords\Exception\ApiException;
@@ -19,7 +20,11 @@ use OCA\Passwords\Services\Object\ShareRevisionService;
 use OCA\Passwords\Services\Object\ShareService;
 use OCP\AppFramework\Http;
 use OCP\AppFramework\Http\JSONResponse;
+use OCP\IConfig;
+use OCP\IGroupManager;
 use OCP\IRequest;
+use OCP\IUser;
+use OCP\IUserManager;
 use OCP\Share\IManager;
 
 /**
@@ -27,36 +32,14 @@ use OCP\Share\IManager;
  *
  * @package OCA\Passwords\Controller\Api
  */
-class ShareApiController extends AbstractObjectApiController {
+class ShareApiController extends AbstractApiController {
+
+    const USER_SEARCH_LIMIT = 512;
 
     /**
-     * @var ShareService
+     * @var IUser
      */
-    protected $modelService;
-
-    /**
-     * @var ShareRevisionService
-     */
-    protected $revisionService;
-
-    /**
-     * @var ShareObjectHelper
-     */
-    protected $objectHelper;
-
-    /**
-     * @var PasswordService
-     */
-    private $passwordModelService;
-
-    /**
-     * @var PasswordRevisionService
-     */
-    private $passwordRevisionService;
-    /**
-     * @var IManager
-     */
-    private $shareManager;
+    protected $user;
 
     /**
      * @var string
@@ -64,96 +47,161 @@ class ShareApiController extends AbstractObjectApiController {
     protected $userId;
 
     /**
-     * @var array
+     * @var IConfig
      */
-    protected $allowedFilterFields = ['created', 'updated', 'cseType', 'sseType'];
+    protected $config;
+
+    /**
+     * @var IUserManager
+     */
+    protected $userManager;
+    /**
+     * @var ShareService
+     */
+    protected $modelService;
+
+    /**
+     * @var IManager
+     */
+    protected $shareManager;
+
+    /**
+     * @var IGroupManager
+     */
+    protected $groupManager;
+
+    /**
+     * @var PasswordService
+     */
+    protected $passwordModelService;
+
+    /**
+     * @var PasswordRevisionService
+     */
+    protected $passwordRevisionService;
 
     /**
      * TagApiController constructor.
      *
      * @param string                  $appName
-     * @param string             $userId
+     * @param IUser                   $user
+     * @param IConfig                 $config
      * @param IRequest                $request
      * @param IManager                $shareManager
+     * @param IUserManager            $userManager
      * @param ShareService            $modelService
-     * @param ShareRevisionService    $revisionService
+     * @param IGroupManager           $groupManager
      * @param PasswordService         $passwordModelService
      * @param PasswordRevisionService $passwordRevisionService
-     * @param ShareObjectHelper       $objectHelper
      */
     public function __construct(
         string $appName,
-        string $userId,
+        IUser $user,
+        IConfig $config,
         IRequest $request,
         IManager $shareManager,
+        IUserManager $userManager,
         ShareService $modelService,
-        ShareRevisionService $revisionService,
+        IGroupManager $groupManager,
         PasswordService $passwordModelService,
-        PasswordRevisionService $passwordRevisionService,
-        ShareObjectHelper $objectHelper
+        PasswordRevisionService $passwordRevisionService
     ) {
-        parent::__construct($appName, $request, $modelService, $revisionService, $objectHelper);
+        parent::__construct(
+            $appName,
+            $request,
+            'PUT, POST, GET, DELETE, PATCH',
+            'Authorization, Content-Type, Accept',
+            1728000
+        );
 
+        $this->user                    = $user;
+        $this->config                  = $config;
+        $this->userId                  = $user->getUID();
+        $this->userManager             = $userManager;
+        $this->groupManager            = $groupManager;
+        $this->modelService            = $modelService;
         $this->shareManager            = $shareManager;
         $this->passwordModelService    = $passwordModelService;
         $this->passwordRevisionService = $passwordRevisionService;
-        $this->userId                  = $userId;
     }
 
     /**
-     * @param string $password
-     * @param string $type
-     * @param string $with
-     * @param bool   $editable
-     * @param string $cseType
+     * @NoCSRFRequired
+     * @NoAdminRequired
      *
-     * @TODO check if share exists for $with user
-     * @TODO check if cseType is same as with all other shares
+     * @param string   $password
+     * @param string   $receiver
+     * @param string   $type
+     * @param int|null $expires
+     * @param bool     $editable
+     * @param bool     $shareable
      *
      * @return JSONResponse
      */
     public function create(
         string $password,
-        string $type,
-        string $with,
+        string $receiver,
+        string $type = 'user',
+        int $expires = null,
         bool $editable = false,
-        string $cseType = EncryptionService::DEFAULT_CSE_ENCRYPTION
+        bool $shareable = false
     ): JSONResponse {
-
         try {
             $this->checkAccessPermissions();
-            $passwordModel = $this->passwordModelService->findByUuid($password);
-            /** @var PasswordRevision $passwordRevision */
-            $passwordRevision = $this->passwordRevisionService->findByUuid($passwordModel->getRevision());
 
-            if($passwordRevision->getCseType() !== EncryptionService::CSE_ENCRYPTION_NONE) {
-                throw new ApiException('Resource has CSE activated', 420);
+            $partners = $this->getSharePartners('');
+            if(!isset($partners[ $receiver ])) throw new ApiException('Invalid receiver uid', 403);
+
+            if(empty($expires)) $expires = null;
+            if($expires !== null && $expires < time()) {
+                throw new ApiException('Invalid expiration date', 400);
+            }
+            if($type !== 'user') {
+                throw new ApiException('Invalid share type', 400);
             }
 
-            $share = $this->modelService->create(
-                $passwordModel->getUuid(),
-                $type,
-                $with
-            );
+            /** @var Password $model */
+            $model = $this->passwordModelService->findByUuid($password);
+            if($model->getShareId()) {
+                $sourceShare = $this->modelService->findByUuid($model->getShareId());
+                $reSharing   = $this->config->getAppValue('core', 'shareapi_allow_resharing', 'yes') === 'yes';
+                if(!$sourceShare->isShareable() || !$reSharing) {
+                    throw new ApiException('Sharing not allowed', 403);
+                }
+                if(!$sourceShare->isEditable()) $editable = false;
+            }
 
-            $revision = $this->revisionService->create(
-                $share->getUuid(),
-                $passwordRevision->getPassword(),
-                $passwordRevision->getUsername(),
-                $passwordRevision->getUrl(),
-                $passwordRevision->getLabel(),
-                $passwordRevision->getNotes(),
-                $passwordRevision->getHash(),
-                $cseType,
-                $editable
-            );
-            $revision->setSynchronized(false);
+            $shares = $this->modelService->findBySourcePasswordAndReceiver($model->getUuid(), $receiver);
+            if($shares !== null) {
+                throw new ApiException('Password already shared with user', 420);
+            }
 
-            $this->revisionService->save($revision);
-            $this->modelService->setRevision($share, $revision);
+            /** @var PasswordRevision $revision */
+            $revision = $this->passwordRevisionService->findByUuid($model->getRevision(), true);
+
+            if($revision->getCseType() !== EncryptionService::CSE_ENCRYPTION_NONE) {
+                throw new ApiException('CSE type does not support sharing', 420);
+            }
+
+            if($revision->getSseType() !== EncryptionService::SSE_ENCRYPTION_V1) {
+                $revision = $this->passwordRevisionService->clone(
+                    $revision,
+                    ['sseType' => EncryptionService::SSE_ENCRYPTION_V1]
+                );
+                $this->passwordRevisionService->save($revision);
+                $this->passwordModelService->setRevision($model, $revision);
+            }
+
+            $share = $this->modelService->create($model->getUuid(), $receiver, $type, $editable, $expires, $shareable);
+            $this->modelService->save($share);
+
+            if(!$model->hasShares()) {
+                $model->setHasShares(true);
+                $this->passwordModelService->save($model);
+            }
 
             return $this->createJsonResponse(
-                ['id' => $share->getUuid(), 'revision' => $revision->getUuid()], Http::STATUS_CREATED
+                ['id' => $share->getUuid()], Http::STATUS_CREATED
             );
         } catch (\Throwable $e) {
             return $this->createErrorResponse($e);
@@ -161,40 +209,47 @@ class ShareApiController extends AbstractObjectApiController {
     }
 
     /**
-     * @param string $id
-     * @param bool   $editable
-     * @param string $cseType
+     * @NoCSRFRequired
+     * @NoAdminRequired
      *
-     * @TODO check if cseType matches password cseType
-     * @TODO check if cseType is same as with all other shares or update
+     * @param string   $id
+     * @param int|null $expires
+     * @param bool     $editable
+     * @param bool     $shareable
      *
      * @return JSONResponse
      */
-    public function update(
-        string $id,
-        bool $editable = false,
-        string $cseType = EncryptionService::DEFAULT_CSE_ENCRYPTION
-    ): JSONResponse {
+    public function update(string $id, int $expires = null, bool $editable = false, bool $shareable = true): JSONResponse {
 
         try {
             $this->checkAccessPermissions();
+
+            if(empty($expires)) $expires = null;
+            if($expires !== null && $expires < time()) {
+                throw new ApiException('Invalid expiration date', 400);
+            }
+
             $share = $this->modelService->findByUuid($id);
-            /** @var ShareRevision $oldRevision */
-            $oldRevision = $this->revisionService->findByModel($share->getUuid());
+            if($share->getUserId() !== $this->userId) {
+                throw new ApiException('Access denied', 403);
+            }
 
+            $share->setExpires($expires);
+            $share->setEditable($editable);
+            $share->setShareable($shareable);
+            $share->setSourceUpdated(true);
+            $this->modelService->save($share);
 
-            /** @var ShareRevision $newRevision */
-            $newRevision = $this->revisionService->clone($oldRevision, ['editable' => $editable, 'cseType' => $cseType]);
-            $this->revisionService->save($newRevision);
-            $this->modelService->setRevision($share, $newRevision);
-
-            return $this->createJsonResponse(['id' => $share->getUuid(), 'revision' => $newRevision->getUuid()]);
+            return $this->createJsonResponse(['id' => $share->getUuid()]);
         } catch (\Throwable $e) {
             return $this->createErrorResponse($e);
         }
     }
 
     /**
+     * @NoCSRFRequired
+     * @NoAdminRequired
+     *
      * @param string $id
      *
      * @return JSONResponse
@@ -202,7 +257,11 @@ class ShareApiController extends AbstractObjectApiController {
     public function delete(string $id): JSONResponse {
         try {
             $this->checkAccessPermissions();
-            $model       = $this->modelService->findByUuid($id);
+            $model = $this->modelService->findByUuid($id);
+            if($model->getUserId() !== $this->userId) {
+                throw new ApiException('Access denied', 403);
+            }
+
             $this->modelService->delete($model);
 
             return $this->createJsonResponse(['id' => $model->getUuid()]);
@@ -212,41 +271,89 @@ class ShareApiController extends AbstractObjectApiController {
     }
 
     /**
-     * @param string $id
-     * @param null   $revision
+     * @NoCSRFRequired
+     * @NoAdminRequired
      *
      * @return JSONResponse
      */
-    public function restore(string $id, $revision = null): JSONResponse {
+    public function info(): JSONResponse {
         try {
             $this->checkAccessPermissions();
-            $model = $this->modelService->findByUuid($id);
 
-            if($revision === null) $revision = $model->getRevision();
-            $oldRevision = $this->revisionService->findByUuid($revision);
+            $info = [
+                'enabled'   => $this->shareManager->shareApiEnabled() &&
+                               !$this->shareManager->sharingDisabledForUser($this->userId),
+                'resharing' => $this->config->getAppValue('core', 'shareapi_allow_resharing', 'yes') === 'yes',
+                'types'     => ['user']
+            ];
 
-            if($oldRevision->getModel() !== $model->getUuid()) {
-                throw new ApiException('Invalid revision id', 400);
-            }
-
-            /** @var ShareRevision $newRevision */
-            $newRevision = $this->revisionService->clone($oldRevision);
-            $this->revisionService->save($newRevision);
-            $this->modelService->setRevision($model, $newRevision);
-
-            return $this->createJsonResponse(['id' => $model->getUuid(), 'revision' => $newRevision->getUuid()]);
+            return $this->createJsonResponse($info);
         } catch (\Throwable $e) {
             return $this->createErrorResponse($e);
         }
     }
 
     /**
+     * @NoCSRFRequired
+     * @NoAdminRequired
+     *
+     * @param string $search
+     *
+     * @return JSONResponse
+     */
+    public function partners(string $search = ''): JSONResponse {
+        try {
+            $this->checkAccessPermissions();
+
+            $partners = [];
+            if($this->config->getAppValue('core', 'shareapi_allow_share_dialog_user_enumeration', 'no') === 'yes') {
+                $partners = $this->getSharePartners($search);
+            }
+
+            return $this->createJsonResponse($partners);
+        } catch (\Throwable $e) {
+            return $this->createErrorResponse($e);
+        }
+    }
+
+    /**
+     * @param string $pattern
+     *
+     * @return array
+     */
+    protected function getSharePartners(string $pattern): array {
+        $partners = [];
+        if($this->shareManager->shareWithGroupMembersOnly()) {
+            $userGroups = $this->groupManager->getUserGroupIds($this->user);
+            foreach ($userGroups as $userGroup) {
+                $users = $this->groupManager->displayNamesInGroup($userGroup, $pattern, self::USER_SEARCH_LIMIT);
+                foreach ($users as $uid => $name) {
+                    if($uid == $this->userId) continue;
+                    $partners[ $uid ] = $name;
+                }
+                if(count($partners) >= self::USER_SEARCH_LIMIT) break;
+            }
+        } else {
+            $usersTmp = $this->userManager->search($pattern, self::USER_SEARCH_LIMIT);
+
+            foreach ($usersTmp as $user) {
+                if($user->getUID() == $this->userId) continue;
+                $partners[ $user->getUID() ] = $user->getDisplayName();
+            }
+        }
+
+        return $partners;
+    }
+
+    /**
      * @throws ApiException
      */
     protected function checkAccessPermissions(): void {
-        $this->shareManager->shareApiEnabled();
-        if($this->shareManager->sharingDisabledForUser($this->userId)) {
+        if(!$this->shareManager->shareApiEnabled()) {
             throw new ApiException('Sharing disabled', 403);
+        }
+        if($this->shareManager->sharingDisabledForUser($this->userId)) {
+            throw new ApiException('Sharing disabled for user', 403);
         }
 
         parent::checkAccessPermissions();
